requirements.txt
Flask==3.1.0
requests==2.32.3
openai==1.77.0
pandas==2.2.2
faiss-cpu=1.11.0
dotenv==0.9.9
python-docx==1.1.2